pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "customer-service-agent"
version = "0.1.0"
description = "A modular customer service agent with LangChain, FastAPI, and Qdrant."
authors = [
    { name = "User", email = "user@example.com" }
]
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "langchain>=0.1.10",
    "langchain-openai>=0.1.0",
    "langchain-groq>=0.1.0",
    "langchain-community>=0.0.25",
    "langchain-ollama>=0.1.0",
    "mem0ai>=0.0.12",
    "streamlit>=1.30.0",
    "langgraph>=0.0.26",
    "langgraph-checkpoint-postgres>=1.0.1",
    "psycopg[pool]",
    "requests>=2.31.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
    "neo4j>=5.0.0",
]
requires-python = ">=3.11"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.2.0",
    "isort>=5.13.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
